% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-lives-in-county.R, R/widget-lives-in-msa.R, R/widget-lives-in-zip.R, R/widget-works-in-county.R, R/widget-works-in-msa.R, R/widget-works-in-zip.R
\name{address}
\alias{address}
\alias{lives_in_county}
\alias{lives_in_msa}
\alias{lives_in_zip}
\alias{works_in_county}
\alias{works_in_msa}
\alias{works_in_zip}
\title{Address based widgets}
\usage{
lives_in_county(..., type = "H")

lives_in_msa(..., type = "H")

lives_in_zip(..., type = "H")

works_in_county(..., type = "B")

works_in_msa(..., type = "B")

works_in_zip(..., type = "B")
}
\arguments{
\item{...}{the counties/msas/zips}

\item{type}{The address type code included (eg "H" for "Home")}
}
\value{
A definition of type \code{entity_id}
}
\description{
By default the "lives in" widgets use an address type code of "H" (Home)
and the "works in" widgets use "B" (Business). Use the \code{type} argument
to change included address types.
}
\examples{
## find anyone who lives or works in the san francisco MSA
lives_in_msa(san_francisco) \%or\% works_in_msa(san_francisco)

## don't forget about synonym search if you're not sure about a code
lives_in_county(?alameda)

## find people who live in any of these counties
lives_in_county(alameda_ca, san_mateo_ca)

}

