% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-age-between.R
\name{age_between}
\alias{age_between}
\title{Age widget}
\usage{
age_between(from = NULL, to = NULL)
}
\arguments{
\item{from}{Minimum age, as a number}

\item{to}{Maximum age, as a number}
}
\description{
Find entities based on age
}
\details{
If an entity has a birth date, then that will be used to calculate their age.
If there is no birth date, then the degree year of their first degree is used.
In order to calculate age from degree year, the widget adds 22 to the number of
years since an undergraduate degree, and 25 to the number of years since a
graduate degree. Entities without a birthday or degree year are completely
ignored.

`from` and `to` are optional, but you must provide at least one of them.
}
\examples{
## entities between the age of 35 and 40
age_between(from = 35, to = 40)

## entities who are exactly 37 years old
age_between(from = 37, to = 37)

## anyone over the age of 55
age_between(from = 55)

## anyone under the age of 65
age_between(to = 65)

}
