% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bots-helpers.R
\name{as_code}
\alias{as_code}
\alias{bot_as_code}
\title{Turn Bot results into valid disco-engine code}
\usage{
as_code(bot_results)

bot_as_code(bot_results)
}
\arguments{
\item{bot_results}{Results from a disco engine bot, such as
\code{\link{brainstorm_bot}} or \code{\link{matrix_bot}}.}
}
\description{
This is a useful helper function for when you want to use several of the
results from the \code{\link{braistorm_bot}} or \code{\link{matrix_bot}} as
part of a Disco Engine definition.
}
\examples{
bb = brainstorm_bot("robot*")
bot_as_code(bb)

}
