% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miner-asset-miner.R
\name{asset_miner}
\alias{asset_miner}
\title{Find entities based on keyword searches of their wealth and assets}
\usage{
asset_miner(..., author = NULL)
}
\arguments{
\item{...}{Search string(s)}

\item{author}{Optionally, include an entity ID of a prospect researcher to only
look at notes written by that researcher.}
}
\value{
A discoveryengine list definition of type \code{entity_id}
}
\description{
Searches through wealth and assets for the given keywords/phrases, and returns
the relevant entities. Searches the "Description" of an asset along with the
"Comment." Enter one
or more search strings. If you enter multiple search strings, the search will
be for notes that contain any one of the searches. Wildcards (*)
are allowed at the beginning or end of each search term (but not in the
middle). For advanced searches, use the \code{ora}
function to use an arbitrary Oracle-style regex as the search term.
}
\examples{
## broad search for those with a trust as a listed asset
asset_miner("trust")

}
