% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-attended-event.R
\name{attended_event}
\alias{attended_event}
\title{Event attendance widget}
\usage{
attended_event(..., participation = c("P", "ST", "SP", "V", "H", "S", "C",
  "KN", "MD", "E"))
}
\arguments{
\item{...}{event codes}

\item{participation}{event participation codes}
}
\description{
Event attendance widget
}

