% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-attended-hs.R
\name{attended_hs}
\alias{attended_hs}
\title{Attended high school widget}
\usage{
attended_hs(...)
}
\arguments{
\item{...}{high-school institution codes}
}
\value{
A discoveryengine list definition of type \code{entity_id}
}
\description{
Find entities who attended a particular high school
}
\examples{
attended_hs(050656)

}
