% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-ca-gave-to-candidate.R,
%   R/widget-ca-gave-to-proposition.R, R/widget-ca-gave.R
\name{ca_campaign}
\alias{ca_campaign}
\alias{ca_gave_to_candidate}
\alias{ca_gave_to_proposition}
\alias{ca_gave}
\title{California campaign finance disclosure widgets}
\usage{
ca_gave_to_candidate(..., at_least = 0.01, from = NULL, to = NULL)

ca_gave_to_proposition(..., at_least = 0.01, from = NULL, to = NULL,
  support = NULL)

ca_gave(at_least = 0.01, from = NULL, to = NULL)
}
\arguments{
\item{...}{Candidate ID(s) or ballot proposition IDs (see examples)}

\item{at_least}{Minimum of aggregate contributions}

\item{from}{begin and end dates (contributed between those dates). Enter as integer of the form YYYYMMDD}

\item{to}{begin and end dates (contributed between those dates). Enter as integer of the form YYYYMMDD}

\item{support}{TRUE/FALSE, whether to look for supporters (TRUE) or those opposed (FALSE) to the ballot initiative. Defaults to both (support and oppose)}
}
\description{
Find entities who have disclosed political contributions in California
statewide election campaigns. \code{ca_gave_to_candidate} only allows you to
search for donors to candidates in statewide races (excluding candidates in local
races), while \code{ca_gave_to_proposition} allows you to search for donors to
ballot initatives. In order to search for all types of giving in California
elections (including local races such as superior court judge or city council),
use \code{ca_gave}
}
\details{
Since this is based on a probabilistic match score, there will be some
false positive matches. If perfect accuracy is necessary, be sure to manually
review the results. Use the built-in lookup features to identify the correct
IDs. For instance, \code{ca_gave_to_proposition(?bag)} will help find the
correct ID for Prop 65, the charge on carry-out bags.
}
\examples{
## example: find donors to Kamala Harris's Attorney General campaign
## first find her candidate ID
ca_gave_to_candidate(?kamala)

## now use the ID:
ca_gave_to_candidate(CA770144)

## look for donors who opposed marijuana decriminialization
ca_gave_to_proposition(?marijuana)

## and use the resulting ID:
## looking only for supporters:
ca_gave_to_proposition(BL20150064, support = TRUE)

}
