% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishing.R
\name{check_out}
\alias{check_out}
\title{Check out a single}
\usage{
check_out(...)
}
\arguments{
\item{...}{The name of the single to check out}
}
\description{
Check out a previously published single. Use \code{\link{show_singles}} to
find the single you're interested in, and check it out by name. In order to
use a single you've checked out, you'll probably want to assign it a name
(see examples).
}
\details{
Note that \code{check_out} will return a single, but will not save it
in your current working environment. So in order to actually use the single,
you should assign it a name (see examples)
}
\examples{
\dontrun{
## note how i assign the single a name (it can be the same as the published
## name, but doesn't have to be).
## i'll check out the in_bay_area single (found by browsing using show_singles)
in_bay_area = check_out(in_bay_area)

## the name i assign does not have to be the same as the published name:
bay = check_out(in_bay_area)

## use the checked out single like you'd use any other definition you created
## example: display
display(bay)

## example: combine with other widgets
bay_math = bay \%and\% majored_in(mathematics)
}
}
