% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-citizen-of.R, R/widget-has-visa.R
\name{citizenship}
\alias{citizenship}
\alias{citizen_of}
\alias{has_visa}
\title{Citizenship/Visa widgets}
\usage{
citizen_of(...)

has_visa(...)
}
\arguments{
\item{...}{country code(s) (for \code{citizen_of}) or visa codes (for \code{has_visa})}
}
\value{
A definition of type \code{entity_id}
}
\description{
These widgets are based on what is found in the "Biographical Detail
(Birth & Death)" screen in CADSbase. More recent student/alumni often have
a Visa code recorded (including cases where the status is "US Citizen"), but
most other records have no information in this field. Be aware that the data
selectively missing.
}
\examples{
## find citizens of Japan
citizen_of(JAPAN)

## find people with a student visa
has_visa(F1)

}
