% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-class-campaign-year.R
\name{class_campaign_year}
\alias{class_campaign_year}
\title{Reunion year widget}
\usage{
class_campaign_year(...)
}
\arguments{
\item{...}{reunion year(s) (eg: 1980, 1990:2000)}
}
\description{
Reunion year widget
}

