% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-contact-text-contains.R
\name{contact_text_contains}
\alias{contact_text_contains}
\title{Find contact reports that contain specific keywords/phrases}
\usage{
contact_text_contains(...)
}
\arguments{
\item{...}{One or more search terms, in quotation marks}
}
\value{
A discoveryengine list definition of type \code{contact_report_id}
}
\description{
This widget is comparable to the Contact Report Miner in CADSMart. Enter one
or more search strings. If you enter multiple search strings, the search will
be for contact reports that contain any one of the searches. Wildcards (*)
are allowed at the beginning or end of each search term (but not in the
middle). This widget creates a definition of type \code{contact_report_id}.
To find the relevant contacted entities, use
\code{\link{contacted_entity_of}}. For advanced searches, use the \code{ora}
function to use an arbitrary Oracle-style regex as the search term.
}
\examples{
## contact reports related to neuroscience
contact_text_contains("neuroscience")

## but what about neurobiology, neuroeconomics, etc?
contact_text_contains("neuro*")

## can use as many search terms as you want. widget will search for
## contact reports containing any one of the search terms
contact_text_contains("neuro*", "brains")

}
\seealso{
\code{\link{contacted_entity_of}}, \code{\link{contact_reports}}
}
