% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter-corp-parent.R,
%   R/converter-corp-subsidiary.R
\name{corp_hierarchy}
\alias{corp_hierarchy}
\alias{corp_parent}
\alias{corp_subsidiary}
\title{Get the parent/subsidiary corporation(s) associated with corporate entity ID(s)}
\usage{
corp_parent(...)

corp_subsidiary(...)
}
\arguments{
\item{...}{one or more entity IDs, or disco engine definitions of type \code{entity_id}}
}
\description{
Every corporate record is included as its own own parent and subsidiary,
including standalone corporate records (where there is no additional hierarchy).
Otherwise, parent/subsidiary records will match what you see in the
corporate hierarchy screen in CADS. Person records do not have corporate
parents or subsidiaries.
}
\note{
If you want to \code{\link{display}} a definition that includes
corporate records, be sure to include the argument
\code{include_organizations = TRUE}, because organization records are
excluded by default (along with deceased records) when using \code{\link{display}}
(see examples).
}
\examples{
# general mills is the parent record of annie's homegrown:
corp_parent(3038697)

# can enter multiple IDs
corp_subsidiary(2090340, 2000407)

# can also enter other definitions
# this just gets me back to 2090340
google = corp_parent(corp_subsidiary(2090340))

# note that i have to use include_organizations = TRUE when displaying
display(google, include_organizations = TRUE)

}
\seealso{
\code{\link{works_at}}
}
