% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-fec-gave-to-candidate.R,
%   R/widget-fec-gave-to-committee.R
\name{fec_gave_to_candidate}
\alias{fec_gave_to_candidate}
\alias{fec}
\alias{fec_gave_to_committee}
\title{FEC disclosure widgets}
\usage{
fec_gave_to_candidate(..., at_least = 0.01, from = NULL, to = NULL)

fec_gave_to_committee(..., at_least = 0.01, from = NULL, to = NULL)
}
\arguments{
\item{...}{FEC Candidate or Committee code(s)}

\item{at_least}{Minimum of total contributions}

\item{from}{begin and end dates (contributed between those dates). Enter as integer of the form YYYYMMDD}

\item{to}{begin and end dates (contributed between those dates). Enter as integer of the form YYYYMMDD}
}
\value{
A definitio of type \code{entity_id}
}
\description{
Find entities who have disclosed political contributions in a federal election.
Data is based on a match of CADS entities to FEC disclosures.
}
\details{
Since this is based on a probabilistic match score, there will be some
false positive matches. If perfect accuracy is necessary, be sure to manually
review the results. Also, FEC disclosure forms show giving to committees that are
registered with the FEC. In order to determine giving to a particular candidate,
we use a crosswalk linking candidates to the committees that support them.
In some instances, a committee will support multiple candidates, so for
example a search for contributors to Kamala Harris's senate campaign will
include in the results any contributors to the committee "Off the Sidelines Senate 2016",
which, in addition to Kamala Harris, also supported Tammy Duckworth and Catherine Masto.
Furthermore, some political donors will give to a PAC that then turns around and
gives to a different committee, and somewhere down the line they support a political
candidate. These donors may not be found when searching using \code{fec_gave_to_candidate}.
}
