% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishing.R
\name{find_singles}
\alias{find_singles}
\alias{single_for}
\title{Search for a particular single using keywords}
\usage{
find_singles(search_term)

single_for(search_term)
}
\arguments{
\item{search_term}{A string to search for in the singles database}
}
\description{
Search for a particular single using keywords
}
\seealso{
\code{\link{show_singles}} for an interactive browser,
\code{\link{check_out}} to check out a single
}
