% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-fund-area.R
\name{fund_area}
\alias{fund_area}
\title{Find funds with a particular area of giving code}
\usage{
fund_area(...)
}
\value{
A discoveryengine list definition of type \code{allocation_code}
}
\description{
This widget does not look for donors to specific types of funds, but rather
creates a definition of type \code{allocation_code} of funds with a particular
area of giving.
}
\examples{
## looking for funds supporting diversity in CED
ced_diversity_fund = fund_text_contains("diversity") \%and\%
    fund_area(environmental_design)

## see the allocation codes:
display(ced_diversity_fund)

## find 1K+ donors:
gave_to_fund(ced_diversity_fund, at_least = 1000)

}
\seealso{
\code{\link{fund_department}}, \code{\link{gave_to_area}}, \code{\link{gave_to_fund}}
}
