% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-fund-department.R
\name{fund_department}
\alias{fund_department}
\title{Find funds with a particular department code}
\usage{
fund_department(...)
}
\value{
A discoveryengine list definition of type \code{allocation_code}
}
\description{
This widget does not look for donors to specific types of funds, but rather
creates a definition of type \code{allocation_code} of funds with a particular
department.
}
\examples{
## are there any PACS funds related to Mahatma Gandhi?
pacs_gandhi_fund = fund_text_contains("gandhi") \%and\%
    fund_department(peace_and_conflict_studies)

## see the allocation codes:
display(pacs_gandhi_fund)

## find 1K+ donors:
gave_to_fund(pacs_gandhi_fund, at_least = 1000)

}
\seealso{
\code{\link{fund_area}}, \code{\link{gave_to_area}}, \code{\link{gave_to_fund}}
}
