% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miner-fund-miner.R
\name{fund_miner}
\alias{fund_miner}
\title{Find donors to funds by keyword/phrase search on the funds}
\usage{
fund_miner(..., at_least = 0.01, from = NULL, to = NULL)
}
\arguments{
\item{...}{Search string(s)}

\item{at_least}{minimum giving to the selected funds}

\item{from}{(optional) start date for looking at giving}

\item{to}{(optional) end date for giving}
}
\description{
This widget is equivalent to running \code{\link{fund_text_contains}} and
\code{\link{gave_to_fund}}.
}
\seealso{
\code{\link{gave_to_fund}}, \code{\link{fund_text_contains}}
}

