% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-fund-purpose.R
\name{fund_purpose}
\alias{fund_purpose}
\title{Find funds with a particular fund purpose}
\usage{
fund_purpose(...)
}
\value{
A discoveryengine list definition of type \code{allocation_code}
}
\description{
This widget does not look for donors to specific types of funds, but rather
creates a definition of type \code{allocation_code} of funds with the specified
fund purpose code(s), e.g. 'Undergrad Student Support' (undergrad scholarships).
Use \code{\link{gave_to_fund}} to find donors to particular funds or types of funds
}
\examples{
## looking for scholarship funds
schol_fund = fund_purpose(USS, GSS, BSS)

## see the allocation codes:
display(schol_fund)

## find 100K+ donors:
gave_to_fund(schol_fund, at_least = 100000)

}
\seealso{
\code{\link{fund_department}}, \code{\link{fund_area}},
\code{\link{fund_type}}, \code{\link{gave_to_fund}},
\code{\link{gave_to_purpose}}
}
