% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-fund-text-contains.R
\name{fund_text_contains}
\alias{fund_text_contains}
\title{Find funds by keyword/phrase search}
\usage{
fund_text_contains(...)
}
\value{
A discoveryengine list definition of type \code{allocation_code}
}
\description{
This widget is comparable to the Fundfinder report in CADSMart. It will
search the fund terms, fund biography, and name of every fund. Enter one
or more search strings. If you enter multiple search strings, the search will
be for funds that contain any one of the searches. Wildcards (*)
are allowed at the beginning or end of each search term (but not in the
middle). This widget creates a definition of type \code{allocation_code}.
To find the donors to a list of funds, use
\code{\link{gave_to_fund}}. For advanced searches, use the \code{ora}
function to use an arbitrary Oracle-style regex as the search term.
}
\examples{
## find funds supporting diversity
diversity_funds = fund_text_contains("divers*", "underrepresented")

## donors to those funds
gave_to_fund(diversity_funds)

## just donors over $10K to those funds
gave_to_fund(diversity_funds, at_least = 10000)

## just donors over $10K during calendar year 2015
gave_to_fund(diversity_funds, at_least = 10000, from = 20150101, to = 20151231)

## donors who had given at least $100K to diversity funds before 2012
gave_to_fund(diversity_funds, at_least = 100000, to = 20111231)

}
\seealso{
\code{\link{gave_to_fund}}
}

