% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-fund-type.R
\name{fund_type}
\alias{fund_type}
\title{Find funds with a particular fund type code}
\usage{
fund_type(...)
}
\value{
A discoveryengine list definition of type \code{allocation_code}
}
\description{
This widget does not look for donors to specific types of funds, but rather
creates a definition of type \code{allocation_code} of funds with the specified
fund type code(s), e.g. capital or endowment
}
\examples{
## looking for capital funds
capital_fund = fund_type(capital)

## see the allocation codes:
display(capital_fund)

## find 100K+ donors:
gave_to_fund(capital_fund, at_least = 100000)

}
\seealso{
\code{\link{fund_department}}, \code{\link{fund_area}}
\code{\link{gave_to_area}}, \code{\link{gave_to_fund}}
}
