% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gave_to_area.R
\name{gave_to_area}
\alias{gave_to_area}
\alias{gave_to_area_}
\title{Area of giving widget
Find entities who have given to specific areas}
\usage{
gave_to_area(..., atleast = 0.01, from = NULL, to = NULL,
  env = parent.frame())

gave_to_area_(aogs, atleast = 0.01, from = NULL, to = NULL,
  env = parent.frame())
}
\arguments{
\item{...}{Area(s) of giving}

\item{atleast}{minimum total giving}

\item{from, }{to begin and end dates (gave between those dates)}
}
\description{
Area of giving widget
Find entities who have given to specific areas
}

