% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-gave-to-department.R
\name{gave_to_department}
\alias{gave_to_department}
\title{Department of giving widget
Find entities who have given to specific departments}
\usage{
gave_to_department(..., at_least = 0.01, from = NULL, to = NULL)
}
\arguments{
\item{...}{Department(s) of giving}

\item{at_least}{minimum total giving}

\item{from, }{to begin and end dates (gave between those dates)}
}
\description{
Department of giving widget
Find entities who have given to specific departments
}

