% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter-gave-to-fund.R
\name{gave_to_fund}
\alias{gave_to_fund}
\title{Giving to fund(s)}
\usage{
gave_to_fund(savedlist, at_least = 0.01, from = NULL, to = NULL)
}
\arguments{
\item{savedlist}{A list/definition of allocation codes}

\item{at_least}{minimum total giving}

\item{from}{begin and end dates (gave between those dates). Enter as an integer of the form YYYYMMDD}

\item{to}{begin and end dates (gave between those dates). Enter as an integer of the form YYYYMMDD}
}
\description{
Find entities who have given to a custom list of funds
}
\examples{
## find funds supporting diversity
diversity_funds = fund_text_contains("diversity")

## donors to those funds
gave_to_fund(diversity_funds)

## just donors over $10K to those funds
gave_to_fund(diversity_funds, at_least = 10000)

}
\seealso{
\code{\link{gave_to_area}}, \code{\link{gave_to_department}}
}

