% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-gave-to-area.R, R/widget-gave-to-department.R
\name{giving}
\alias{gave_to_area}
\alias{gave_to_department}
\alias{giving}
\title{Giving widgets}
\usage{
gave_to_area(..., at_least = 0.01, from = NULL, to = NULL)

gave_to_department(..., at_least = 0.01, from = NULL, to = NULL)
}
\arguments{
\item{...}{Area(s)/Department(s) of giving}

\item{at_least}{minimum total giving}

\item{from}{begin and end dates (gave between those dates). Enter as an integer of the form YYYYMMDD}

\item{to}{begin and end dates (gave between those dates). Enter as an integer of the form YYYYMMDD}
}
\value{
A definition of type \code{entity_id}
}
\description{
Find entities who have given to specific types of funds
}
\details{
giving is calculated on a pledged basis, and includes soft credits.
For dates, the giving record date is used. If no area/department is entered,
widget will find donors to anything.
}
\examples{
## gave at least $5,000 to cal performances during FY 2016
gave_to_area(cal_performances, at_least = 5000, from = 20150701, to = 20161231)

## has given at least $100,000 lifetime to engineering
gave_to_area(engineering, at_least = 100000)

}
\seealso{
\code{\link{gave_to_fund}}
}

