% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-has-occupation.R
\name{has_occupation}
\alias{has_occupation}
\title{Find people with a given occupation_code}
\usage{
has_occupation(..., include_former = FALSE)
}
\arguments{
\item{...}{Occupation code(s)}

\item{include_former}{Should non-current employment data be included in the
search? If FALSE (the default) only current employment will be searched}
}
\value{
A definition of type \code{entity_id}
}
\description{
This widget is based on the "Occupational Spec" fields on an entity's
Employment screen.
}

