% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-has-philanthropic-affinity.R
\name{has_philanthropic_affinity}
\alias{has_philanthropic_affinity}
\title{Philanthropic affinity widget}
\usage{
has_philanthropic_affinity(..., at_least = 0)
}
\arguments{
\item{...}{Philanthropic affinity type code(s)}

\item{at_least}{Minimum total giving to the selected philanthropic affinity type(s)}
}
\value{
A definition of type \code{entity_id}
}
\description{
Philanthropic affinity widget
}
\details{
This widget allows you to search by the Org Type, and not by specific
philanthropic organizations. Use \code{at_least} to find just those who have
given over a certain amount to such organizations. \code{at_least} uses total
giving to the selected org types. If no codes are entered,
widget will find anyone with any philanthropic affinity.
}
\examples{
has_philanthropic_affinity(health_medicine)

## people who have given at least $10K TOTAL to
## human rights and social welfare organizations
has_philanthropic_affinity(human_rights, social_welfare, at_least = 10000)

}
