% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-has-philanthropic-interest.R
\name{has_philanthropic_interest}
\alias{has_philanthropic_interest}
\title{Philanthropic interest widget}
\usage{
has_philanthropic_interest(..., unit = NULL, active = TRUE,
  comment = NULL)
}
\arguments{
\item{...}{UCB Interest Codes or synonyms}

\item{unit}{Look only for philanthropic interests associated with a given unit.
Leave blank (the default) to search only on phil. interest codes}

\item{active}{If TRUE (default), will include only philanthropic interests
without a stop date. If FALSE, will include only philanthropic interests
WITH a stop date -- the stop date is used here as a disqualifier, to signify negative
interest/inclination.}

\item{comment}{(Optional) Supply one or more search terms to search through
the comment fields of the philanthropic interests area}
}
\value{
A definition of type \code{entity_id}
}
\description{
Find entities with a given philanthropic interests.
}
\examples{
## find people with an interest in funding scholarships
has_philanthropic_interest(scholarships)

## can also be used to exclude disqualified prospects
donor = gave_to_area(at_least = 100000)
disqualified = has_philanthropic_interest(UC, active = FALSE)
donor \%but_not\% disqualified

}
