% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-has-position.R
\name{has_position}
\alias{has_position}
\title{Find people who have a specific position level code}
\usage{
has_position(..., include_former = FALSE)
}
\arguments{
\item{...}{position level code(s). If left empty, will look for any
non-blank position level code}

\item{include_former}{Should non-current employment data be included in the
search? If FALSE (the default) only current employment will be searched}
}
\value{
A definition of type \code{entity_id}
}
\description{
This widget is based on the "position level" code in CADS, which is entered
by Prospect Research and identifies Director-level and above employment
positions
}

