% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-has-record-type.R
\name{has_record_type}
\alias{has_record_type}
\title{Find entities with a given record type(s)}
\usage{
has_record_type(...)
}
\arguments{
\item{...}{record type code(s)}
}
\value{
A definition of type \code{entity_id}
}
\description{
This widget looks at all record types (not just the primary record type).
If no record types are entered, the widget will find every entity ID,
regardless of type. However, please pay attention to the default settings
on \code{\link{display}}, which excludes organizations and the deceased
unless you specify otherwise.
}
\examples{
## find trustees
has_record_type(ucbf_trustee)

## or if you want to include former trustees
has_record_type(ucbf_trustee, former_ucbf_trustee)

}

