% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-is-high-net-worth-parent.R
\name{is_high_net_worth_parent}
\alias{is_high_net_worth_parent}
\title{Select entities that have been coded as "High Net Worth" Parents}
\usage{
is_high_net_worth_parent()
}
\description{
Select entities that have been coded as "High Net Worth" Parents
}
\details{
No arguments necessary; this function calls all entities with an active code
of "High Net Worth Parent"
}
