% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter-household.R
\name{household}
\alias{household}
\title{Household a list of entities}
\usage{
household(savedlist)
}
\arguments{
\item{savedlist}{an existing discoveryengine definition}
}
\description{
The \code{\link{display}} function has a built-in option to household, so
you will usually not need to use this function.
}
\details{
Will return just the primary ID for a household, instead of both spouses. Note
that it will not household additional parts of the definition that are added
after the `household` command (see examples).
}
\examples{
is_wealthy = has_capacity(1:7)
is_wealthy_hh = household(is_wealthy)

## this will NOT household the additional part of the definition
is_wealthy_hh \%or\% lives_in_msa(san_francisco)

## better to get everything together, then household:
household(is_wealthy \%or\% lives_in_msa(san_francisco))

}
