% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-in-savedlist.R
\name{in_savedlist}
\alias{in_savedlist}
\title{Find entities in a given savedlist}
\usage{
in_savedlist(...)
}
\arguments{
\item{...}{Savedlist ID(s) -- see \code{\link{show_savedlists}}}
}
\value{
A definition of type \code{entity_id}
}
\description{
To find the ID of a savedlist, use \code{\link{show_savedlists}}.
}
\seealso{
\code{\link{show_savedlists}}
}
