% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-in-suspect-pool.R
\name{in_suspect_pool}
\alias{in_suspect_pool}
\title{Find entities in a given suspect pool}
\usage{
in_suspect_pool(...)
}
\arguments{
\item{...}{Suspect Pool ID(s) -- see \code{\link{show_suspect_pools}}}
}
\value{
A definition of type \code{entity_id}
}
\description{
The disco engine creates unique, persistent IDs for the suspect pools, in
order to make them queryable. To find the ID of your suspect pool, use
\code{\link{show_suspect_pools}}.
}
\seealso{
\code{\link{show_suspect_pools}}
}

