% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-in-unit-portfolio.R
\name{in_unit_portfolio}
\alias{in_unit_portfolio}
\title{Entities in a unit portfolio}
\usage{
in_unit_portfolio(..., include_inactive = FALSE)
}
\arguments{
\item{...}{Office code(s)}

\item{include_inactive}{Include inactive proposals/assignments?}
}
\description{
Entities in a unit portfolio
}

