% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-is-a.R
\name{is_a}
\alias{is_a}
\title{Specify entity type and status}
\usage{
is_a(..., include_deceased = FALSE)
}
\arguments{
\item{...}{entity type (person, organization)}

\item{include_deceased}{Should deceased individuals and out of business
corporations be included? (See details)}
}
\description{
The \code{\link{display}} function has built-in options to include/exclude
organizations and deceased or out of business corporations, so you will
usually not need to use this function.
}
\details{
By default, deceased individuals and out of business corporations are
excluded.
}
