% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-job-title-like.R
\name{job_title_like}
\alias{job_title_like}
\title{Find entities based on keyword searches of their job titles}
\usage{
job_title_like(..., include_former = FALSE)
}
\arguments{
\item{...}{Search string(s)}

\item{include_former}{Should non-current employment data be included in the
search? If FALSE (the default) only current employment will be searched}
}
\value{
A discoveryengine list definition of type \code{entity_id}
}
\description{
\code{\link{has_occupation}}, \code{\link{works_in_industry}}, and
\code{\link{has_position}} are all based on specific employment codes.
This widget, on the other hand, searches the text of job titles in the
employment table, and finds just those that match the search string(s). As
with any other text-searching widget, search strings must appear within
quotation marks.
If you enter multiple search strings, the search will
be for notes that contain any one of the searches. Wildcards (*)
are allowed at the beginning or end of each search term (but not in the
middle). For advanced searches, use the \code{ora}
function to use an arbitrary Oracle-style regex as the search term.
}
\examples{
## search for jobs that include the word "data"
job_title_like("data")

## the previous search would not include jobs with the word "database",
## but by adding an asterisk, we can capture those jobs also:
job_title_like("data*")

## a search for data analysts:
job_title_like("data*", "analy*")

}
