% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-majored-in.R
\name{majored_in}
\alias{majored_in}
\title{Academic major widget}
\usage{
majored_in(..., undergraduates = TRUE, graduates = TRUE,
  attendees = FALSE)
}
\arguments{
\item{...}{major codes or major code synonyms}

\item{undergraduates}{TRUE/FALSE: should include undergraduates? Default is TRUE}

\item{graduates}{TRUE/FALSE: should include graduates? Default is TRUE}

\item{attendees}{TRUE/FALSE: should include attendees (TRUE) or just degreeholders (FALSE). Default is FALSE}
}
\description{
Find entities who graduated with a particular major(s).
}

