% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter-married-to.R
\name{married_to}
\alias{married_to}
\title{The spouse of a person}
\usage{
married_to(...)
}
\arguments{
\item{...}{one or more existing discoveryengine definition(s), and/or individual entity id(s)}
}
\description{
The spouse of a person
}
\examples{
## looking for CNR donors who are not degreeholders,
## but this doesn't work, because it returns the spouses
## of CNR degreeholders who receive credit for household gifts
gave_to_area(CNR) \%but_not\% has_degree_from(NR)

## want to also exclude anyone married to a CNR degreeholder:
cnr_degreeholder = has_degree_from(NR)
gave_to_area(CNR) \%but_not\%
    (cnr_degreeholder \%or\% married_to(cnr_degreeholder))

}
\seealso{
\code{\link{household}}
}
