% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bots-matrix-bot.R
\name{matrix_bot}
\alias{matrix_bot}
\title{Broaden a search using co-occurrence analysis}
\usage{
matrix_bot(savedlist)
}
\arguments{
\item{savedlist}{A disco engine definition}
}
\description{
The \code{matrix_bot} takes a disco engine definition and compares it to
other codes in CADS, looking for patterns of co-occurrence. That is, it
analyzes individuals who match your definition, looking for other codes they
have in common.
}
\examples{
## we start out with a basic definition of interst in Chicano Studies
chicano_interest = majored_in(chicano_studies) \%or\% gave_to_department(CHICANO)

## now the matrix_bot discovers related characteristics, such as
## majoring in ethnic studies or participating in MEChXA
matrix_bot(chicano_interest)

}

