% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-participated-in.R
\name{participated_in}
\alias{participated_in}
\title{Student Activity widget}
\usage{
participated_in(..., leaders_only = FALSE)
}
\arguments{
\item{...}{Student activity codes/synonyms}

\item{leaders_only}{Pull only leaders of student groups (TRUE) or all
participants (FALSE)? Defaults to FALSE (all participants)}
}
\value{
A definition of type \code{entity_id}
}
\description{
Student Activity widget
}
\details{
If no student activities are entered, then widget will look for
people who participated in any student activity.
}
\examples{
## Nick once needed to find parents of band members who graduated in
## the last 5 years. so, start with recently graduated band members:
recent_band_members = participated_in(MSMB) \%and\% has_reunion_year(2010:2015)

## then get their parents
recent_band_parents = parent_of(recent_band_members)

display(recent_band_parents)

}
