% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people.R
\name{people_}
\alias{people}
\alias{people_}
\title{Basic entity query}
\usage{
people_(.dots, env = parent.frame())

people(..., env = parent.frame())
}
\arguments{
\item{env}{evaluation environment}

\item{...}{any conditions}

\item{include_deceased}{should deceased individuals be included? (Defaults to FALSE)}
}
\description{
Basic entity query
}

