% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-played-sport.R
\name{played_sport}
\alias{played_sport}
\title{Find entities who played a sport}
\usage{
played_sport(...)
}
\arguments{
\item{...}{Sport codes/synonyms}
}
\value{
A definition of type \code{entity_id}
}
\description{
Make sure you know which sport codes to use. For example, men's and women's
basketball have different codes. Take advantage of the synonym search feature
if you're not sure (for example: \code{played_sport(?basketball)}). If no
sports are entered, widget will find people who played any sport (including
intramurals).
}
\examples{
## anyone who played basketball for Cal
played_sport(basketball_men, basketball_women)

}

