% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-in-unit-portfolio.R
\name{portfolio}
\alias{in_unit_portfolio}
\alias{portfolio}
\title{Portfolio widgets}
\usage{
in_unit_portfolio(..., include_inactive = FALSE)
}
\arguments{
\item{...}{Office code(s)}

\item{include_inactive}{Should include inactive proposals/assignments? Defaults to FALSE}
}
\description{
Find entities in a given portfolio
}

