% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-has-implied-capacity.R, R/widget-has-major-gift-score.R
\name{has_implied_capacity}
\alias{has_implied_capacity}
\alias{has_major_gift_score}
\alias{predictive_model}
\title{Predictive model widgets}
\usage{
has_implied_capacity(...)

has_major_gift_score(...)
}
\arguments{
\item{...}{score descriptors(more, most, etc)}
}
\description{
Find entities who have selected scores on predictive models
}

