% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-proposal-development-officer.R
\name{proposal_development_officer}
\alias{proposal_development_officer}
\title{Find proposals assigned to a given development officer}
\usage{
proposal_development_officer(..., include_inactive = TRUE)
}
\arguments{
\item{...}{A discoveryengine definition of type entity_id, and/or individual entity ids}

\item{include_inactive}{whether inactive assignments/proposals should be included}
}
\description{
Find proposals assigned to a given development officer
}
