% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-proposal-office.R
\name{proposal_office}
\alias{proposal_office}
\title{Find proposals assigned to a given unit}
\usage{
proposal_office(..., include_inactive = TRUE)
}
\arguments{
\item{...}{Office code(s)}

\item{include_inactive}{whether inactive assignments/proposals should be included}
}
\description{
Find proposals assigned to a given unit
}
