% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-proposal-purpose.R
\name{proposal_purpose}
\alias{proposal_purpose}
\title{Find proposals with given purpose programs}
\usage{
proposal_purpose(..., target = 0, actual = 0, commit = 0)
}
\arguments{
\item{...}{program code(s)}

\item{target}{purpose target ask amount (at least this much)}

\item{actual}{purpose actual ask amount (at least this much)}

\item{commit}{purpose commit amount (at least this much)}
}
\description{
Find proposals with given purpose programs
}
