% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-proposal-stage.R
\name{proposal_stage}
\alias{proposal_stage}
\title{Find proposals based on current stage}
\usage{
proposal_stage(...)
}
\arguments{
\item{...}{stage code(s)}
}
\description{
Find proposals based on current stage
}
