% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-proposal-disqualified.R,
%   R/widget-proposal-qualified.R, R/widget-proposal-stage-transition.R
\name{proposal_disqualified}
\alias{proposal_disqualified}
\alias{proposal_qualified}
\alias{proposal_stage_transition}
\title{Find proposals that had the given stage transition during a given time period}
\usage{
proposal_disqualified(from = NULL, to = NULL)

proposal_qualified(from = NULL, to = NULL)

proposal_stage_transition(from_stages = NULL, to_stages = NULL,
  from_date = NULL, to_date = NULL)
}
\arguments{
\item{from}{daterange in YYYYMMDD format}

\item{to}{daterange in YYYYMMDD format

Note that \code{proposal_qualified} just looks for transitions between
\code{QU} and any of \code{(CU, SP, PD, GS, DS)}. Similarly,
\code{proposal_disqualified} looks for transitions from
\code{QU} to \code{(TD, DQ)}.}

\item{from_stages}{character vector of one or more stages}

\item{to_stages}{character vector of one or more stages}

\item{from_date}{daterange in YYYYMMDD format}

\item{to_date}{daterange in YYYYMMDD format}
}
\description{
Find proposals that had the given stage transition during a given time period
}
