% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-proposal-type.R
\name{proposal_type}
\alias{proposal_type}
\title{Find proposals based on type}
\usage{
proposal_type(...)
}
\arguments{
\item{...}{proposal type code(s)}
}
\description{
Find proposals based on type
}
