% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-rated-by.R
\name{rated_by}
\alias{rated_by}
\title{Search by capacity rating researcher(s) and/or date(s)}
\usage{
rated_by(..., from = NULL, to = NULL, include_inactive = FALSE)
}
\arguments{
\item{...}{(optional) One or more entity IDs of researcher(s). If no
researchers are entered, then the widget will look for anyone who was rated
(during the given dates).}

\item{from}{(optional) date, Enter as an integer of the form YYYYMMDD}

\item{to}{(optional) date, Enter as an integer of the form YYYYMMDD}

\item{include_inactive}{TRUE or FALSE: Should widget look through just active ratings (FALSE),
or all ratings (TRUE)?}
}
\value{
A definition of type \code{entity_id}
}
\description{
This widget allows you to search for entities who have been rated, by the
researcher who rated them and/or the date range during which the entity
was rated. If no researcher IDs are entered, then the widget will look for
anyone who was rated (or, if \code{from} and \code{to} dates are entered, just
those rated during the specified daterange).
}
\examples{
## find anyone ever rated by Lucila
rated_by(3221525)

## find anyone rated by Lucila since 1/1/2017
rated_by(3221525, from = 20170101)

## find anyone rated by Lucila during calendar year 2016
rated_by(3221525, from = 20160101, to = 20161231)

## find anyone rated by Lucila or Mallory on or before June 30, 2016
## including ratings that are no longer active
rated_by(3221525, 923406, to = 20160630, include_inactive = TRUE)

}
