% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-received-award.R
\name{received_award}
\alias{received_award}
\title{Find entities who've received an award}
\usage{
received_award(..., comment = NULL)
}
\arguments{
\item{...}{award/honor codes}

\item{comment}{(Optional) Supply one or more search terms to search through
the comment field}
}
\description{
Find entities who've received an award
}
\examples{
## find anyone who has won a nobel prize in physics
received_award(nobel_prize, comment = "physics")

}
