% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter-related-to.R
\name{related_to}
\alias{related_to}
\title{A relation of a person}
\usage{
related_to(savedlist, ..., comment = NULL)
}
\arguments{
\item{savedlist}{A discoveryengine definition}

\item{...}{Relationship type codes (\code{relation_type_code}). Leave empty
to search for all relationships.}
}
\description{
For an existing disco engine list definition, get relations of those entities.
Not entering any relationship type codes results in finding all relations
(brothers, sisters, colleagues, spouses, etc.).
}
\details{
Many relationship types in CADS have multiple codes, to represent each side
of the relationship type. For instance, "Mother" is encoded by both "SM"
(presumably for "Son-Mother") and "DM" ("Daughter-Mother"). This can lead to
confusion and frustration, so be sure that you're including all relevant codes.
You can do so by using synonym lookup or just using synonyms in your definition
(see examples).
}
\examples{
## say we want to identify siblings of founders
## start by identifying founders, via the founders pledge
founder = has_affiliation(founders_pledge)

## brother and sister each have two different codes (see details above)
founder_sibling = related_to(founder, brother, sister)
display(founder_sibling)

}
\seealso{
\code{\link{parent_of}}, \code{\link{child_of}}
}
