% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-sec-filed.R
\name{sec}
\alias{sec}
\alias{sec_filed}
\title{Find entities who made SEC filings}
\usage{
sec_filed(
  ...,
  from = NULL,
  to = NULL,
  director = NULL,
  officer = NULL,
  ten_percenter = NULL,
  other = NULL,
  title_text = NULL,
  verified = NULL
)
}
\arguments{
\item{...}{Central Index Key for one or more companies -- if blank, will look at all SEC filings}

\item{from}{begin and end dates (filed between those dates). Enter as integer of the form YYYYMMDD}

\item{to}{begin and end dates (filed between those dates). Enter as integer of the form YYYYMMDD}

\item{director}{Look for people listed as directors? See details}

\item{officer}{Look for people listed as officers? See details}

\item{ten_percenter}{Look for people listed as ten percent owners? See details}

\item{other}{Look for people with some other relationship to the company? See details}

\item{title_text}{Search term to look for in filer's free-text entry for officer or other}

\item{verified}{See details}
}
\value{
A definition of type \code{entity_id}
}
\description{
Find entities who have filed an SEC Form 3/4/5. Data is based on a match of
CADS entities to SEC disclosures
}
\details{
By default, all matched filings are included in the search. If
\code{verified} is TRUE, then only those that have been verified by the
Prospect Discovery Team are included. If \code{verified} is FALSE, then only
filings that have NOT been verified by Prospect Discovery are included.
#'
Since this is based on a probabilistic match score, there will be some
false positive matches. If perfect accuracy is necessary, be sure to manually
review the results. You might want to use the `screening` chunk from the
`discoappend` package, which outputs links to filings on the SEC's website.

SEC Form 4 includes checkboxes for "Director," "Officer," "ten percent owner,"
and "Other" to specify the filer's relationship to the company. Furthermore,
filers who check "officer" can enter their title as free text, and those who
check "Other" can similarly enter their title or other explanatory text.
Search these entries using the `title_text` argument.
}
\examples{
## look up companies using the usual look-up feature
sec_filed(?twitter)

## with a code, i can now pull any filers with Twitter:
sec_filed(1418091)

## To look for anyone who's filed during calendar year 2017:
sec_filed(1418091, from = 20170101, to = 20171231)

## find people who are 10\% owners of any company
sec_filed(ten_percenter = TRUE)

## find filers with any company who have listed their title as "chairman"
sec_filed(title_text = "chairman")

}
