% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-savedlists.R
\name{show_savedlists}
\alias{show_savedlists}
\title{Browse all saved lists}
\usage{
show_savedlists()
}
\description{
Pops up an interactive table that allows you to sort and search through a
listing of all saved lists.
}
\seealso{
\code{\link{in_savedlist}}
}
