% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishing.R
\name{list_singles}
\alias{list_singles}
\alias{show_singles}
\title{Browse available singles}
\usage{
list_singles()

show_singles()
}
\description{
Any time a user \code{\link{publish}}es a new single, it gets listed and stored
along with keywords and a description that should make it easier to find.
\code{show_singles} allows you to view singles that have been published and
search through them.
}
\seealso{
\code{\link{publish}}, \code{\link{check_out}}, \code{\link{find_singles}}
}

