% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-suspect-pools.R
\name{show_suspect_pools}
\alias{show_suspect_pools}
\title{Browse all suspect pools}
\usage{
show_suspect_pools()
}
\description{
Pops up an interactive table that allows you to sort and search through a
listing of all suspect pools.
}
\seealso{
\code{\link{in_suspect_pool}}
}

