% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-find-widgets.R
\name{show_widgets}
\alias{show_widgets}
\title{Browse the widget registry}
\usage{
show_widgets()
}
\description{
Pops up an interactive table that allows you to sort and search through a
listing of all available widgets, with descriptions of each
}
\seealso{
\code{\link{widget_for}}
}
