% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-suggestion-bot.R
\name{suggestion_bot}
\alias{suggestion_bot}
\title{Suggest widgets and codes based on a search term}
\usage{
suggestion_bot(search_term)
}
\arguments{
\item{search_term}{a single search term}
}
\description{
The \code{suggestion_bot} takes a search term and searches through all code
tables in CADS. If any of the codes it finds happen to be covered by a
discoveryengine widget, it brings them back and suggests them.
}
\examples{
suggestion_bot("neuroscience")
}

