% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-find-widgets.R
\name{widget_for}
\alias{widget_for}
\title{Look for widgets using keywords}
\usage{
widget_for(search_term = ".*")
}
\arguments{
\item{search_term}{a string to search for in the widget registry}
}
\description{
Look for widgets using keywords
}
\examples{
widget_for("sports")

}
\seealso{
\code{\link{show_widgets}} for an interactive widget browser
}
