% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter-works-at.R
\name{works_at}
\alias{works_at}
\title{Find entities who work for a given employer(s)}
\usage{
works_at(..., include_former = FALSE)
}
\arguments{
\item{...}{one or more entity IDs, or discovery engine definition(s) of type \code{entity_id}}

\item{include_former}{Should former employment be considered (default is FALSE)}
}
\description{
Note that you can specify one or more employers, as plain entity IDs or as
disco engine definitions. See examples.
}
\examples{
# anyone who works at google corporate
works_at(2090340)

# anyone who works at google or any of its subsidiaries
works_at(2090340, corp_subsidiary(2090340))

# works at google or facebook, including any of their subsidiaries
works_at(2090340, 2145589, corp_subsidiary(2090340, 2145589))

}
