% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-works-in-industry.R
\name{works_in_industry}
\alias{works_in_industry}
\title{Find people who work in a particular industry}
\usage{
works_in_industry(..., include_former = FALSE)
}
\arguments{
\item{...}{NAICS code(s)}

\item{include_former}{Should non-current employment data be included in the
search? If FALSE (the default) only current employment will be searched}
}
\value{
A definition of type \code{entity_id}
}
\description{
This widget is based on the 6-digit NAICS code found on the employer's
"Industry Codes" screen.
}
\examples{
## find people who work in investment banking
works_in_industry(523110)

}
