% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-has-degree-from.R, R/widget-majored-in.R
\name{academic}
\alias{academic}
\alias{has_degree_from}
\alias{majored_in}
\title{Academic widgets}
\usage{
has_degree_from(..., undergraduates = TRUE, graduates = TRUE,
  attendees = FALSE)

majored_in(..., undergraduates = TRUE, graduates = TRUE,
  attendees = FALSE)
}
\arguments{
\item{...}{Schools/majors}

\item{undergraduates}{TRUE/FALSE: should include undergraduates? Default is TRUE}

\item{graduates}{TRUE/FALSE: should include graduates? Default is TRUE}

\item{attendees}{TRUE/FALSE: should include attendees (TRUE) or just degreeholders (FALSE). Default is FALSE}
}
\value{
A definition of type \code{entity_id}
}
\description{
Find entities who graduated with a particular major(s) or from a particular
school(s). By default, attendees are not included, use \code{attendees = TRUE}
to include attendees. If no majors/schools are entered, the widget will look
for anyone with a Cal degree (or if \code{attendees = TRUE}, anyone who has
attended Cal). Will not retrieve anyone with non-Berkeley degrees.
}
\examples{
## majored in philosophy and/or math
majored_in(mathematics, philosophy)

## just math grad degree holders
majored_in(mathematics, undergraduates = FALSE)

## with haas, we often want to include attendees because of the evening/wknd program:
has_degree_from(haas, attendees = TRUE)

}
\seealso{
\code{\link{has_reunion_year}}
}

