% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-lives-in-county.R,
%   R/widget-lives-in-foreign-country.R, R/widget-lives-in-msa.R,
%   R/widget-lives-in-state.R, R/widget-lives-in-zip.R,
%   R/widget-works-in-county.R, R/widget-works-in-foreign-country.R,
%   R/widget-works-in-msa.R, R/widget-works-in-state.R, R/widget-works-in-zip.R
\name{address}
\alias{address}
\alias{lives_in_county}
\alias{lives_in_foreign_country}
\alias{lives_in_msa}
\alias{lives_in_state}
\alias{lives_in_zip}
\alias{works_in_county}
\alias{works_in_foreign_country}
\alias{works_in_msa}
\alias{works_in_state}
\alias{works_in_zip}
\title{Address based widgets}
\usage{
lives_in_county(..., include_alternate = TRUE, include_past = FALSE,
  include_seasonal = FALSE, include_student_local = FALSE,
  include_student_permanent = FALSE, type = "home")

lives_in_foreign_country(..., include_alternate = TRUE,
  include_past = FALSE, include_seasonal = FALSE,
  include_student_local = FALSE, include_student_permanent = FALSE,
  city = NULL, type = "home")

lives_in_msa(..., include_alternate = TRUE, include_past = FALSE,
  include_seasonal = FALSE, include_student_local = FALSE,
  include_student_permanent = FALSE, type = "home")

lives_in_state(..., include_alternate = TRUE, include_past = FALSE,
  include_seasonal = FALSE, include_student_local = FALSE,
  include_student_permanent = FALSE, type = "home")

lives_in_zip(..., include_alternate = TRUE, include_past = FALSE,
  include_seasonal = FALSE, include_student_local = FALSE,
  include_student_permanent = FALSE, type = "home")

works_in_county(..., include_alternate = TRUE, include_past = FALSE,
  include_self_employed = TRUE, type = "business")

works_in_foreign_country(..., include_alternate = TRUE,
  include_past = FALSE, include_self_employed = TRUE, city = NULL,
  type = "business")

works_in_msa(..., include_alternate = TRUE, include_past = FALSE,
  include_self_employed = TRUE, type = "business")

works_in_state(..., include_alternate = TRUE, include_past = FALSE,
  include_self_employed = TRUE, type = "business")

works_in_zip(..., include_alternate = TRUE, include_past = FALSE,
  include_self_employed = TRUE, type = "business")
}
\arguments{
\item{...}{the counties/msas/zips}

\item{include_alternate}{include addresses coded as "alternate?" Defaults to TRUE}

\item{include_past}{include past/former addresses? Defaults to FALSE}

\item{include_seasonal}{for lives_in_*: include seasonal addresses? Defaults to FALSE}

\item{include_student_local}{for lives_in_*: include student local address? Defaults to FALSE}

\item{include_student_permanent}{for lives_in_*: include student permanent address? Defaults to FALSE}

\item{type}{Will be either "home" or "business". Advanced users can specify individual address type codes (see examples)}

\item{city}{For foreign addresses, a search string to find individual cities (see details and examples)}

\item{include_self_employed}{for works_in_*: include self-employed business addresses? Defaults to TRUE}
}
\value{
A definition of type \code{entity_id}
}
\description{
By default the "lives in" widgets use an address type code of "H" (Home)
and the "works in" widgets use "B" (Business). Use the \code{type} argument
to change included address types.
}
\details{
Foreign addresses do not have associated county or MSA codes, and
currently (as of March 2018) are not geocoded in CADS, so will not be found to
the \code{\link{lives_near}} and \code{\link{works_near}} widgets. There are
no strict code tables associated with cities, but the \code{city} argument in
the foreign country widgets allow you to search for a particular city by name
(see examples)
}
\examples{
## find anyone who lives or works in the san francisco MSA
lives_in_msa(san_francisco) \%or\% works_in_msa(san_francisco)

## don't forget about synonym search if you're not sure about a code
lives_in_county(?alameda)

## find people who live in any of these counties
lives_in_county(alameda_ca, san_mateo_ca)

## the best way to find people working in Seoul
works_in_foreign_country(KORER, city = "Seoul")

## people who live abroad or used to live abroad:
lives_in_foreign_country(include_past = TRUE)

## you shouldn't need to, but if necessary
## you can directly specify address type codes:
## make sure you know what you're doing!
lives_in_msa(san_francisco, type = "S")

}
