% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attended_event.R
\name{attended_event}
\alias{attended_event}
\alias{attended_event_}
\title{Event attendance widget}
\usage{
attended_event(..., participation = c("P", "ST", "SP", "V", "H", "S", "C",
  "KN", "MD", "E"), env = parent.frame())

attended_event_(events, participation = c("P", "ST", "SP", "V", "H", "S", "C",
  "KN", "MD", "E"), env = parent.frame())
}
\arguments{
\item{...}{event codes}

\item{participation}{event participation codes}
}
\description{
Event attendance widget
}

