% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-attended-event.R
\name{attended_event}
\alias{attended_event}
\title{Event attendance widget}
\usage{
attended_event(..., from = NULL, to = NULL, include_non_attendees = FALSE,
  comment = NULL)
}
\arguments{
\item{...}{event codes}

\item{from}{begin and end dates (attended event between those dates). Enter as an integer of the form YYYYMMDD}

\item{to}{begin and end dates (attended event between those dates). Enter as an integer of the form YYYYMMDD}

\item{include_non_attendees}{do you want to include people who were invited,
but did not attend (regrets, no-show, invited but did not attend) defaults to FALSE}

\item{comment}{(Optional) Supply one or more search terms to search through
the comment field}
}
\value{
A discoveryengine list definition of type \code{entity_id}
}
\description{
Find entities who attended (an) event(s).
}
\details{
Enter one or more event codes, and decide whether you want to include people
who were on the invite list but did not attend (by default, they are excluded).
If you don't enter any event codes, then you will get people who attended any
event.
}
\examples{
attended_event(3965, 1263)

}
