% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-attended-event.R
\name{attended_event}
\alias{attended_event}
\title{Event attendance widget}
\usage{
attended_event(..., include_non_attendees = FALSE)
}
\arguments{
\item{...}{event codes}

\item{include_non_attendees}{do you want to include people who were invited,
but did not attend (regrets, no-show, invited but did not attend) defaults to FALSE}
}
\value{
A discoveryengine list definition of type \code{entity_id}
}
\description{
Find entities who attended (an) event(s).
}
\details{
Enter one or more event codes, and decide whether you want to include people
who were on the invite list but did not attend (by default, they are excluded).
If you don't enter any event codes, then you will get people who attended any
event.
}
\examples{
attended_event(3965, 1263)

}

