% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bots-brainstorm-bot.R
\name{brainstorm_bot}
\alias{brainstorm_bot}
\title{Suggest widgets and codes based on a search term}
\usage{
brainstorm_bot(search_term)
}
\arguments{
\item{search_term}{a single search term}
}
\description{
The \code{brainstorm_bot} takes a search term and searches through all code
tables in CADS. If any of the codes it finds happen to be covered by a
discoveryengine widget, it brings them back and suggests them.
}
\examples{
brainstorm_bot("neuroscience")
}

