% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-gave-to-area.R
\name{gave_to_area}
\alias{gave_to_area}
\title{Area of giving widget
Find entities who have given to specific areas}
\usage{
gave_to_area(..., at_least = 0.01, from = NULL, to = NULL)
}
\arguments{
\item{...}{Area(s) of giving}

\item{from, }{to begin and end dates (gave between those dates)}

\item{atleast}{minimum total giving}
}
\description{
Area of giving widget
Find entities who have given to specific areas
}

