% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter-gave-to-fund.R
\name{gave_to_fund}
\alias{gave_to_fund}
\title{Giving to fund(s)}
\usage{
gave_to_fund(savedlist, at_least = 0.01, from = NULL, to = NULL)
}
\arguments{
\item{savedlist}{A list/definition of allocation codes}

\item{at_least}{minimum total giving (defaults to .01)}

\item{from}{begin and end dates (gave between those dates). Enter as an integer of the form YYYYMMDD}

\item{to}{begin and end dates (gave between those dates). Enter as an integer of the form YYYYMMDD}
}
\value{
A discoveryengine list definition of type \code{entity_id}
}
\description{
Find entities who have given to a custom list of funds
}
\details{
\code{from} and \code{to} are optional, not entering them will leave that end
of the date range open. So for example a year-to-date cutoff can be achieved
by using \code{from} but not \code{to}. Leaving both blank will look for
donors at any time. Giving is aggregated to all funds in \code{savedlist} --
so you're not looking for someone who gave \code{at_least} to one of the funds
on the list, but rather someone who gave in total \code{at_least} to all/any
of the funds.
}
\examples{
## find funds supporting diversity
diversity_funds = fund_text_contains("diversity")

## donors to those funds
gave_to_fund(diversity_funds)

## just donors over $10K to those funds
gave_to_fund(diversity_funds, at_least = 10000)

## just donors over $10K during calendar year 2015
gave_to_fund(diversity_funds, at_least = 10000, from = 20150101, to = 20151231)

## donors who had given at least $100K to diversity funds before 2012
gave_to_fund(diversity_funds, at_least = 100000, to = 20111231)

}
\seealso{
\code{\link{gave_to_area}}, \code{\link{gave_to_department}}, \code{\link{fund_text_contains}}
}
