% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-has-affiliation.R
\name{has_affiliation}
\alias{has_affiliation}
\title{Affiliation widget}
\usage{
has_affiliation(..., include_former = TRUE, comment = NULL)
}
\arguments{
\item{...}{affiliation code(s)}

\item{include_former}{TRUE/FALSE should include former affiliations? Defaults to TRUE}

\item{comment}{(Optional) Supply one or more search terms to search through
the comment field}
}
\value{
A definition of type \code{entity_id}
}
\description{
Find entities with a given affiliation(s). If no affiliation codes are entered,
widget will find entities who have any affiliation code.
}
\examples{
## let's find some SF elites
has_affiliation(san_francisco_grid_club, bohemian_club)

## look for just current homecoming volunteers
has_affiliation(caa_homecoming_volunteer, include_former = FALSE)

}
