% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_affiliation.R
\name{has_affiliation}
\alias{has_affiliation}
\alias{has_affiliation_}
\title{Affiliation widget}
\usage{
has_affiliation(..., status = c("C", "F"), env = parent.frame())

has_affiliation_(affiliations, status = c("C", "F"), env = parent.frame())
}
\arguments{
\item{...}{affiliation code(s)}

\item{status}{affilition status codes ("C" for current, "F" for former)}
}
\description{
Affiliation widget
}

