% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-has-affiliation.R
\name{has_affiliation}
\alias{has_affiliation}
\title{Affiliation widget}
\usage{
has_affiliation(..., include_former = TRUE)
}
\arguments{
\item{...}{affiliation code(s)}

\item{status}{affilition status codes ("C" for current, "F" for former)}
}
\description{
Affiliation widget
}

