% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-has-implied-capacity.R
\name{has_implied_capacity}
\alias{has_implied_capacity}
\alias{has_implied_capacity_}
\title{Implied capacity widget}
\usage{
has_implied_capacity(...)

has_implied_capacity_(likelihood)
}
\arguments{
\item{...}{Predictive model score descriptions (most, more, etc)}
}
\description{
Implied capacity widget
}

