% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-has-interest.R
\name{has_interest}
\alias{has_interest}
\title{Interest code widget}
\usage{
has_interest(
  ...,
  include_children = FALSE,
  include_former = FALSE,
  comment = NULL
)
}
\arguments{
\item{...}{Interest code(s)}

\item{include_children}{If TRUE, will include interest codes in the interest group listed
(for example, when searching "health", if TRUE, will include interest codes
such as Aging, Nutritional Science, Parkinson's Disease, etc.).}

\item{include_former}{If TRUE, will include all interest codes,
even if they have a stop date. If FALSE (default), will exclude codes with a stop date.}

\item{comment}{(Optional) Supply one or more search terms to search through
the comment fields of the interest area}
}
\value{
A definition of type \code{entity_id}
}
\description{
Find entities with a given interest(s)
}
\details{
If no interests are entered, widget will search for entities with any interest
code.
}
\examples{
has_interest(data_science)

## prospects for cancer research can be interest coded with "cancer"
## or with the more general "health"
has_interest(cancer) \%or\% has_interest(health, comment = "cancer")

}
