% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-has-interest.R
\name{has_interest}
\alias{has_interest}
\title{Interest code widget}
\usage{
has_interest(..., include_former = TRUE)
}
\arguments{
\item{...}{Interest code(s)}

\item{include_former}{If TRUE (default), will include all interest codes,
even if they have a stop date. If FALSE, will exclude codes with a stop date.}
}
\value{
A definition of type \code{entity_id}
}
\description{
Find entities with a given interest(s)
}
\details{
If no interests are entered, widget will search for entities with any interest
code.
}
\examples{
has_interest(data_science)

## same as
has_interest(DAT)

}
