% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-has-philanthropic-affinity.R
\name{has_philanthropic_affinity}
\alias{has_philanthropic_affinity}
\title{Philanthropic affinity widget}
\usage{
has_philanthropic_affinity(..., at_least = 0)
}
\arguments{
\item{...}{Philanthropic affinity type code(s)}

\item{at_least}{Minimum total giving to the selected philanthropic affinity type(s)}
}
\description{
Philanthropic affinity widget
}

