% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-has-reunion-year.R
\name{has_reunion_year}
\alias{has_reunion_year}
\title{Reunion year widget}
\usage{
has_reunion_year(...)
}
\arguments{
\item{...}{reunion year(s) (eg: 1980, 1990:2000)}
}
\description{
Reunion year widget
}

