% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-on-committee.R
\name{on_committee}
\alias{on_committee}
\title{Committee participation widget}
\usage{
on_committee(..., include_former = TRUE)
}
\arguments{
\item{...}{Committee code(s)}

\item{include_former}{Should former committee memberships be included (TRUE), or just current ones (FALSE)? Defaults to TRUE}
}
\description{
Find entities on a particular committee(s)
}

