% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter-child-of.R, R/converter-parent-of.R
\name{child_of}
\alias{child_of}
\alias{parent_child}
\alias{parent_of}
\title{The parent or child of a person}
\usage{
child_of(...)

parent_of(...)
}
\arguments{
\item{...}{A discoveryengine definition of type \code{entity_id}, and/or individual entity ids}
}
\description{
This uses just "Son" and "Daughter" for the definition of child, and just
"Mother" and "Father" for the definition of parent. For a more custom use of
relationship codes, use \code{\link{related_to}}.
}
\examples{
## finding the children of wealthy parents
is_wealthy = has_capacity(1:9) \%or\% has_implied_capacity(more, most)
child_of(is_wealthy)

## or the parents of wealthy children
parent_of(is_wealthy)
}
\seealso{
\code{\link{related_to}}
}
