% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-in-development-officer-portfolio.R,
%   R/widget-in-unit-portfolio.R
\name{in_development_officer_portfolio}
\alias{in_development_officer_portfolio}
\alias{portfolio}
\alias{in_unit_portfolio}
\title{Portfolio widgets}
\usage{
in_development_officer_portfolio(..., include_inactive = FALSE)

in_unit_portfolio(..., include_inactive = FALSE)
}
\arguments{
\item{...}{Office code(s)}

\item{include_inactive}{Should include inactive proposals/assignments? Defaults to FALSE}
}
\value{
A definition of type \code{entity_id}
}
\description{
Find entities in a given portfolio
}
\details{
Uses the office associated with the proposal assignment to define portfolios.
If no offices are entered, then anyone in any portfolio will be included.
}
\examples{
in_unit_portfolio(governmental_studies, greater_good_science_center)

}
