% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-has-cnr-score.R,
%   R/widget-has-engineering-score.R, R/widget-has-gift-planning-score.R,
%   R/widget-has-haas-score.R, R/widget-has-implied-capacity.R,
%   R/widget-has-major-gift-score.R
\name{has_cnr_score}
\alias{has_cnr_score}
\alias{has_engineering_score}
\alias{has_gift_planning_score}
\alias{has_haas_score}
\alias{has_implied_capacity}
\alias{predictive_model}
\alias{has_major_gift_score}
\title{Predictive model widgets}
\usage{
has_cnr_score(...)

has_engineering_score(...)

has_gift_planning_score(...)

has_haas_score(...)

has_implied_capacity(...)

has_major_gift_score(...)
}
\arguments{
\item{...}{score descriptors(more, most, etc)}
}
\value{
A definition of type \code{entity_id}
}
\description{
Find entities who have selected scores on predictive models
}
\details{
Use score descriptions (more_likely, most_likely, etc). If no score descriptions
are entered, widget will look for anyone with a score on the predictive model.
}
\examples{
has_implied_capacity(more_likely, most_likely, somewhat_likely)
has_major_gift_score(most_likely, more_likely)

}
