% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-proposal-actual-ask.R
\name{proposal_actual_ask}
\alias{proposal_actual_ask}
\title{Find proposals with given ask dates/amounts}
\usage{
proposal_actual_ask(at_least = 0, from = NULL, to = NULL)
}
\arguments{
\item{at_least}{only look for asks of at least this much}

\item{from}{daterange in YYYYMMDD format}

\item{to}{daterange in YYYYMMDD format}
}
\description{
Find proposals with given ask dates/amounts
}
