% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-proposal-disqualified.R,
%   R/widget-proposal-qualified.R, R/widget-proposal-stage-transition.R
\name{proposal_disqualified}
\alias{proposal_disqualified}
\alias{proposal_qualified}
\alias{proposal_stage_transition}
\title{Find proposals that had the given stage transition during a given time period}
\usage{
proposal_disqualified(from = NULL, to = NULL)

proposal_qualified(from = NULL, to = NULL)

proposal_stage_transition(from_stages = NULL, to_stages = NULL,
  from_date = NULL, to_date = NULL)
}
\arguments{
\item{from}{daterange in YYYYMMDD format}

\item{to}{daterange in YYYYMMDD format}

\item{from_stages}{character vector of one or more stages}

\item{to_stages}{character vector of one or more stages}

\item{from_date}{daterange in YYYYMMDD format}

\item{to_date}{daterange in YYYYMMDD format}
}
\description{
\code{proposal_qualified} just looks for transitions between
\code{QU} and any of \code{(CU, SP, PD, GS, DS)}. Similarly,
\code{proposal_disqualified} looks for transitions from
\code{QU} to \code{(TD, DQ)}. \code{proposal_stage_transition} allows you
to specify other types of transitions that might be of interest.
}
\examples{
## successful QUs during FY18
proposal_qualified(from = 20170701, to = 20180630)

## this would do the same thing
proposal_stage_transition(from_stages = "QU",
                          to_stages = c("CU", "SP", "PD", "GS", "DS"),
                          from_date = 20170701, to_date = 20180630)

## from_stages and to_stages are optional, so to find
## proposals that were *ever* in QU:
proposal_stage_transition(from_stages = "QU",
                          from_date = 20170701, to_date = 20180630)

}
