% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-lives-near.R, R/widget-works-near.R
\name{radius}
\alias{radius}
\alias{lives_near}
\alias{works_near}
\title{Geographic radius search}
\usage{
lives_near(location, miles, latitude = NULL, longitude = NULL,
  include_alternate = TRUE, include_past = FALSE,
  include_seasonal = FALSE, include_student_local = FALSE,
  include_student_permanent = FALSE, type = "home")

works_near(location, miles, latitude = NULL, longitude = NULL,
  include_alternate = TRUE, include_past = FALSE,
  include_self_employed = TRUE, type = "business")
}
\arguments{
\item{location}{A (quoted) string describing a place (see examples)}

\item{miles}{The radius of the search, all addresses within this many
miles of \code{location} will be returned}

\item{latitude}{If you already know the lat/long of your search location,
then you can enter them directly rather than entering an address in \code{location}}

\item{longitude}{(see above)}

\item{include_past}{include past/former addresses? Defaults to FALSE}

\item{include_seasonal}{for lives_near: include seasonal addresses? Defaults to FALSE}

\item{include_student_local}{for lives_near: include student local address? Defaults to FALSE}

\item{include_student_permanent}{for lives_near: include student permanent address? Defaults to FALSE}

\item{type}{Will be either "home" or "business". Advanced users can specify individual address type codes}

\item{include_self_employed}{for works_near: include self-employed business addresses? Defaults to TRUE}
}
\value{
A disco definition of type \code{entity_id}
}
\description{
Use this widget to find people who live or work "near" a given location, for
instance to aid with event planning or when a development officer is about to
take a trip.
}
\examples{
## find everyone within 10 miles of campus
lives_near("UC Berkeley", miles = 10)

## can always use exact addresses
lives_near("1995 University Ave., Berkeley, CA 94720", miles = 15.5)

## if you already know the lat/long, you can enter them directly:
works_near(lat = 37.87238, lon = -122.2542, miles = 10)

}
\seealso{
\code{\link{address}}
}
