% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter-related-to.R
\name{related_to}
\alias{related_to}
\title{A relation of a person}
\usage{
related_to(savedlist, ...)
}
\arguments{
\item{savedlist}{A discoveryengine definition}

\item{...}{Relationship type codes (\code{relation_type_code})}
}
\description{
A relation of a person
}
\seealso{
\code{\link{parent_of}}, \code{\link{child_of}}
}

